Revert "Make some of the SDF parser tests more generic"
