--- multibody/multibody_tree/parsing/parser_common.cc
@@ -1,3 +1,0 @@
-#include "drake/multibody/multibody_tree/parsing/parser_common.h"
-
-// This is an empty file to confirm that our header parses on its own.
